Fix double nested loop bug , bump version -> 0.0.8
I had an error in my scope handling code that would cause an
error on any code following a dbl nested loop. We now only
evaluate a backref if the depth of its scope is back to the same
as the calling level. And if the scope happens to drop below
then we discard the last ref as it's an extra tacked on by the
fact that it takes a final look at the original call before leaving
the loop. (I know, I know, I should be using an AST, blah blah)